requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.2
langchain-huggingface==1.2.0
langchain-openai==1.1.5
langchain-text-splitters==1.1.0
langsmith==0.5.0
markupsafe==3.0.3
marshmallow==3.26.1
modelscope==1.33.0
mpmath==1.3.0
multidict==6.7.0
mypy-extensions==1.1.0
networkx==3.6.1
numpy==2.3.5
openai==2.13.0
orjson==3.11.5
packaging==25.0
propcache==0.4.1
protobuf==3.20.3
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
sentencepiece==0.2.1
setuptools==80.9.0
sniffio==1.3.1
sqlalchemy==2.0.45
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.6.2
uuid-utils==0.12.0
yarl==1.22.0
zstandard==0.25.0
